requirements.txt
faiss-cpu==1.13.2
fastapi==0.133.1
httpx==0.28.1
langchain==1.2.10
langchain-community==0.4.1
langchain-huggingface==1.2.0
langchain-mistralai==1.1.1
mistralai==1.12.4
numpy==2.4.2
pandas==3.0.1
pytest==9.0.2
python-dotenv==1.2.1
sentence-transformers==5.2.3
uvicorn==0.41.0
requests==2.31.0